# encoding/binary : add float support to fast path