encoding/binary: add float support to fast path

This adds float type support to the main switch blocks in Read and
Write, instead of falling back to reflection.  This gives a considerable
speedup for the float types.

It's not clear to me why other tests show other speedups and some slowdowns.

(If anyone can help me understand the slowdown on unrelated functions, I would appreciate it)

```
name                      old time/op    new time/op     delta
ReadSlice1000Int32s-8       4.48µs ±15%     4.84µs ±15%      ~     (p=0.073 n=7+7)
ReadStruct-8                1.10µs ±12%     1.13µs ±13%      ~     (p=0.702 n=6+7)
ReadInts-8                   267ns ±14%      266ns ± 8%      ~     (p=1.000 n=6+7)
WriteInts-8                  322ns ±10%      306ns ± 6%      ~     (p=0.189 n=6+7)
WriteSlice1000Int32s-8      5.00µs ±30%     4.73µs ± 4%      ~     (p=0.927 n=7+7)
PutUint16-8                 0.69ns ±14%     0.64ns ±10%      ~     (p=0.091 n=7+7)
PutUint32-8                 0.69ns ±11%     0.66ns ±10%      ~     (p=0.273 n=7+7)
PutUint64-8                 0.75ns ± 6%     0.76ns ± 2%      ~     (p=0.223 n=7+6)
LittleEndianPutUint16-8     0.50ns ±14%     0.63ns ± 6%   +25.83%  (p=0.001 n=7+7)
LittleEndianPutUint32-8     0.62ns ± 7%     0.48ns ± 3%   -23.41%  (p=0.001 n=7+6)
LittleEndianPutUint64-8     0.49ns ± 5%     0.62ns ± 6%   +24.94%  (p=0.001 n=6+7)
ReadFloats-8                 129ns ± 9%       70ns ± 8%   -46.02%  (p=0.001 n=7+7)
WriteFloats-8                131ns ± 6%       86ns ±11%   -34.59%  (p=0.001 n=7+7)
ReadSlice1000Float32s-8     14.6µs ±14%      4.8µs ±12%   -67.29%  (p=0.001 n=7+7)
WriteSlice1000Float32s-8    16.4µs ±20%      4.7µs ± 8%   -71.01%  (p=0.001 n=7+7)
PutUvarint32-8              24.2ns ±10%     24.9ns ± 7%      ~     (p=0.271 n=7+7)
PutUvarint64-8              64.2ns ± 9%     64.1ns ±15%      ~     (p=0.738 n=7+7)
[Geo mean]                  57.1ns          47.1ns        -17.59%
```